refactor(localize): extract the editor's box-drawing stage into a hook
