+++ frontend/src/components/localize/editor/LocalizeObjectEditor.tsx
@@ -39,18 +39,7 @@ import {
 import { buildFilmstripEntries, type FilmstripEntry } from '@/utils/annotation/objectFilmstrip';
 import { computeCellCrop } from '@/utils/annotation/gridCropUtils';
 import type { AlertFrame } from '@/utils/annotation/alertLocalizeUtils';
-import {
-  calculateImageBounds,
-  screenToImageCoordinates,
-  imageToNormalizedCoordinates,
-  normalizedToImageCoordinates,
-  moveBox,
-  resizeBox,
-  type CurrentDrawing,
-  type ImageBounds,
-  type Point,
-  type ResizeHandle,
-} from '@/utils/annotation';
+import { useBoxDrawingStage } from '@/hooks/annotation';
 import {
   FADE_MS,
   prefersReducedMotion,
@@ -223,27 +212,6 @@ export function LocalizeObjectEditor({
     };
   }, []);
 
-  const [imageInfo, setImageInfo] = useState<{
-    width: number;
-    height: number;
-    offsetX: number;
-    offsetY: number;
-  } | null>(null);
-
-  // Zoom / pan — lifted from ImageModal unchanged; that plumbing is sound.
-  const [zoomLevel, setZoomLevel] = useState(1);
-  const [panOffset, setPanOffset] = useState<Point>({ x: 0, y: 0 });
-  const [transformOrigin, setTransformOrigin] = useState<Point>({ x: 50, y: 50 });
-  const [isDragging, setIsDragging] = useState(false);
-  const [dragStart, setDragStart] = useState({ x: 0, y: 0 });
-
-  const [currentDrawing, setCurrentDrawing] = useState<CurrentDrawing | null>(null);
-  // Space swaps the drag from drawing to panning, as it does in every other
-  // image tool. Mirrored into a ref because mousedown reads it in the same
-  // tick the keydown may have set it.
-  const [spaceHeld, setSpaceHeld] = useState(false);
-  const spaceHeldRef = useRef(false);
-
   // `G` cycles what the idle stage draws: the default pick, every candidate
   // at once, or nothing at all — the bare plume.
   const [boxVisibility, setBoxVisibility] = useState<BoxVisibility>('pick');
@@ -271,15 +239,6 @@ export function LocalizeObjectEditor({
   // what reveals the move/resize affordances.
   const [boxSelected, setBoxSelected] = useState(false);
 
-  const [boxEdit, setBoxEdit] = useState<{
-    startClient: { x: number; y: number };
-    orig: [number, number, number, number];
-    mode: 'move' | 'resize';
-    handle?: ResizeHandle;
-    next: [number, number, number, number];
-  } | null>(null);
-  const didDragBoxRef = useRef(false);
-
   const { data: imageData } = useDetectionImage(detection.id);
   const queryClient = useQueryClient();
 
@@ -295,6 +254,33 @@ export function LocalizeObjectEditor({
     [detection, existingAnnotation]
   );
   const committed = useMemo(() => committedBox(existingAnnotation), [existingAnnotation]);
+
+  // Image geometry, zoom/pan and the modeless pointer handling, shared with
+  // the add-object flow. `commitDrawn` is defined below and reaches the hook
+  // through a ref — the same trick `clearRef` uses for the keyboard handler,
+  // so nothing has to be reordered to satisfy the dependency.
+  const commitDrawnRef = useRef<(xyxyn: [number, number, number, number]) => void>(() => undefined);
+  const stage = useBoxDrawingStage({
+    containerRef,
+    imgRef,
+    // The ORIGINAL box, not the live-dragged one: a drag can only start when
+    // `boxEdit` is null, at which point the two are identical anyway.
+    editableBox: committed?.xyxyn ?? null,
+    boxSelected,
+    onBoxSelectedChange: setBoxSelected,
+    onDrawn: xyxyn => commitDrawnRef.current(xyxyn),
+  });
+  // Destructured because effects depend on them: these are useCallback-stable,
+  // so a plain identifier keeps the dependency arrays honest where `stage.x`
+  // would drag the whole (per-render) object in.
+  const {
+    boxEdit,
+    currentDrawing,
+    resetZoom: resetStageZoom,
+    applyView: applyStageView,
+    resetTransient: resetStageTransient,
+  } = stage;
+
   // A live drag renders from `boxEdit.next` so the box tracks the cursor
   // before the save round-trips.
   const shownCommitted: BoxCandidate | null = boxEdit
@@ -374,6 +360,8 @@ export function LocalizeObjectEditor({
     [peeked, detection, smokeType, onCommit, onCommitGapFrame]
   );
 
+  commitDrawnRef.current = commitDrawn;
+
   const clear = useCallback(() => {
     setPreviewed(null);
     // A frame with no model evidence exists only because a human boxed it;
@@ -425,54 +413,24 @@ export function LocalizeObjectEditor({
 
   // --- Zoom ---------------------------------------------------------------
 
-  const resetZoom = useCallback(() => {
-    setZoomLevel(1);
-    setPanOffset({ x: 0, y: 0 });
-    setTransformOrigin({ x: 50, y: 50 });
-  }, []);
-
-  const constrainPan = useCallback(
-    (offset: Point): Point => {
-      if (!imgRef.current || zoomLevel <= 1) return offset;
-      // Layout size, not the transformed rect: the pan applies INSIDE the
-      // scale, so the max offset keeping the image covering its box is
-      // baseSize*(z-1)/(2z).
-      const maxPanX = (imgRef.current.offsetWidth * (zoomLevel - 1)) / (2 * zoomLevel);
-      const maxPanY = (imgRef.current.offsetHeight * (zoomLevel - 1)) / (2 * zoomLevel);
-      return {
-        x: Math.max(-maxPanX, Math.min(maxPanX, offset.x)),
-        y: Math.max(-maxPanY, Math.min(maxPanY, offset.y)),
-      };
-    },
-    [zoomLevel]
-  );
-
-  useEffect(() => {
-    setPanOffset(prev => constrainPan(prev));
-  }, [constrainPan]);
-
   // `Z`: frame the object rather than the landscape. `computeCellCrop` is the
   // cockpit's own crop-mode math, and its output maps straight onto the
   // canvas's existing zoom/transform-origin props — no second render path.
   useEffect(() => {
     if (!cropView) {
-      resetZoom();
+      resetStageZoom();
       return;
     }
     const boxes = committed ? [committed] : candidates;
     if (boxes.length === 0) return;
-    const crop = computeCellCrop(boxes, OBJECT_FRAMING);
-    setZoomLevel(crop.scale);
-    setPanOffset({ x: 0, y: 0 });
-    setTransformOrigin({ x: crop.originX, y: crop.originY });
+    applyStageView(computeCellCrop(boxes, OBJECT_FRAMING));
     // Re-frames on frame change too, since the object moves between frames.
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [cropView, detection.id]);
 
   // Navigating to another frame resets the transient per-frame state.
   useEffect(() => {
-    setCurrentDrawing(null);
-    setBoxEdit(null);
+    resetStageTransient();
     setBoxSelected(false);
     setBoxVisibility('pick');
     setPreviewed(null);
@@ -481,230 +439,12 @@ export function LocalizeObjectEditor({
     // previous geometry stays correct; clearing it unmounted every overlay
     // until the next image fired onLoad, which read as the boxes blinking out
     // on each arrow press. `handleImageLoad` refreshes it either way.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [detection.id]);
-
-  useEffect(() => {
-    const setHeld = (held: boolean) => {
-      spaceHeldRef.current = held;
-      setSpaceHeld(held);
-    };
-    const onKeyDown = (e: KeyboardEvent) => {
-      if (e.code !== 'Space') return;
-      // Space activates a focused button; only claim it elsewhere. The target
-      // is only an Element when something is focused — a keydown dispatched at
-      // the window itself has no `closest`.
-      const target = e.target;
-      if (target instanceof Element && target.closest('button, input, textarea')) return;
-      e.preventDefault();
-      setHeld(true);
-    };
-    const onKeyUp = (e: KeyboardEvent) => {
-      if (e.code === 'Space') setHeld(false);
-    };
-    // A window that loses focus mid-hold never sees the keyup.
-    const onBlur = () => setHeld(false);
-    window.addEventListener('keydown', onKeyDown);
-    window.addEventListener('keyup', onKeyUp);
-    window.addEventListener('blur', onBlur);
-    return () => {
-      window.removeEventListener('keydown', onKeyDown);
-      window.removeEventListener('keyup', onKeyUp);
-      window.removeEventListener('blur', onBlur);
-    };
-  }, []);
-
-  // --- Coordinates --------------------------------------------------------
-
-  const handleImageLoad = () => {
-    const img = imgRef.current;
-    if (!img || !containerRef.current) return;
-    // LAYOUT metrics, not getBoundingClientRect(): the rect is already scaled
-    // by the zoom transform, and the overlays position themselves from these
-    // numbers and then get the same transform applied on top. Measuring the
-    // transformed rect while opening zoomed recorded a box three times too
-    // large and pushed every overlay off the image. `constrainPan` avoids the
-    // same trap for the same reason.
-    setImageInfo({
-      width: img.offsetWidth,
-      height: img.offsetHeight,
-      offsetX: img.offsetLeft,
-      offsetY: img.offsetTop,
-    });
-  };
-
-  const getImageInfo = (): {
-    containerOffset: Point;
-    imageBounds: ImageBounds;
-    transform: { zoomLevel: number; panOffset: Point; transformOrigin: Point };
-  } | null => {
-    if (!imgRef.current || !containerRef.current) return null;
-    const container = containerRef.current;
-    const containerRect = container.getBoundingClientRect();
-    return {
-      containerOffset: { x: containerRect.left, y: containerRect.top },
-      // LAYOUT metrics for the bounds — same trap as handleImageLoad: while
-      // the open/close animation scales the editor root, the rect is the
-      // transformed visual size, and overlays computed from it during that
-      // window keep the garbage geometry after the animation ends.
-      imageBounds: calculateImageBounds({
-        containerWidth: container.offsetWidth,
-        containerHeight: container.offsetHeight,
-        imageNaturalWidth: imgRef.current.naturalWidth,
-        imageNaturalHeight: imgRef.current.naturalHeight,
-      }),
-      transform: { zoomLevel, panOffset, transformOrigin },
-    };
-  };
-
-  const screenToImageCoords = (screenX: number, screenY: number): Point => {
-    const info = getImageInfo();
-    if (!info) return { x: 0, y: 0 };
-    return screenToImageCoordinates(
-      { x: screenX, y: screenY },
-      info.containerOffset,
-      info.imageBounds,
-      info.transform
-    );
-  };
-
-  const imageToNormalized = (imageX: number, imageY: number): Point => {
-    const info = getImageInfo();
-    if (!info) return { x: 0, y: 0 };
-    return imageToNormalizedCoordinates({ x: imageX, y: imageY }, info.imageBounds);
-  };
-
-  const normalizedToImage = (normX: number, normY: number): Point => {
-    const info = getImageInfo();
-    if (!info) return { x: 0, y: 0 };
-    return normalizedToImageCoordinates({ x: normX, y: normY }, info.imageBounds);
-  };
-
-  // --- Pointer ------------------------------------------------------------
-
-  const handleBoxPointerDown = (e: React.MouseEvent) => {
-    if (!shownCommitted) return;
-    e.stopPropagation();
-    // First click selects; only a selected box can be dragged, so a stray
-    // click on it never nudges the annotation.
-    if (!boxSelected) {
-      setBoxSelected(true);
-      return;
-    }
-    didDragBoxRef.current = false;
-    setBoxEdit({
-      mode: 'move',
-      startClient: { x: e.clientX, y: e.clientY },
-      orig: shownCommitted.xyxyn,
-      next: shownCommitted.xyxyn,
-    });
-  };
-
-  const handleHandlePointerDown = (handle: ResizeHandle, e: React.MouseEvent) => {
-    if (!shownCommitted) return;
-    e.stopPropagation();
-    didDragBoxRef.current = false;
-    setBoxEdit({
-      mode: 'resize',
-      handle,
-      startClient: { x: e.clientX, y: e.clientY },
-      orig: shownCommitted.xyxyn,
-      next: shownCommitted.xyxyn,
-    });
-  };
-
-  /**
-   * The canvas has no modes. A press that misses the box starts drawing one;
-   * hold space or press the middle button to pan instead. Nothing to arm,
-   * nothing to leave armed, and no state to misread — which is what made a
-   * two-click draw dangerous once it was always available.
-   */
-  const handleMouseDown = (e: React.MouseEvent) => {
-    didDragBoxRef.current = false;
-
-    const wantsPan = spaceHeldRef.current || e.button === 1;
-    if (wantsPan) {
-      // Middle-press otherwise starts the browser's autoscroll.
-      if (e.button === 1) e.preventDefault();
-      if (zoomLevel > 1) {
-        setIsDragging(true);
-        setDragStart({ x: e.clientX - panOffset.x, y: e.clientY - panOffset.y });
-      }
-      return;
-    }
-
-    if (e.button !== 0) return;
-    const coords = screenToImageCoords(e.clientX, e.clientY);
-    setCurrentDrawing({
-      startX: coords.x,
-      startY: coords.y,
-      currentX: coords.x,
-      currentY: coords.y,
-    });
-  };
-
-  const handleMouseMove = (e: React.MouseEvent) => {
-    if (boxEdit && imgRef.current) {
-      // Screen-px delta over the image's on-screen size: pan- and
-      // origin-invariant, so the box tracks the cursor 1:1 at any zoom.
-      const displayW = imgRef.current.offsetWidth * zoomLevel;
-      const displayH = imgRef.current.offsetHeight * zoomLevel;
-      const dx = (e.clientX - boxEdit.startClient.x) / displayW;
-      const dy = (e.clientY - boxEdit.startClient.y) / displayH;
-      const next =
-        boxEdit.mode === 'move'
-          ? moveBox(boxEdit.orig, dx, dy)
-          : resizeBox(boxEdit.orig, boxEdit.handle as ResizeHandle, dx, dy);
-      didDragBoxRef.current = true;
-      setBoxEdit(prev => (prev ? { ...prev, next } : prev));
-      return;
-    }
-    if (currentDrawing) {
-      const coords = screenToImageCoords(e.clientX, e.clientY);
-      setCurrentDrawing(prev =>
-        prev ? { ...prev, currentX: coords.x, currentY: coords.y } : null
-      );
-    } else if (isDragging && zoomLevel > 1) {
-      setPanOffset(constrainPan({ x: e.clientX - dragStart.x, y: e.clientY - dragStart.y }));
-    }
-  };
-
-  const handleMouseUp = () => {
-    if (boxEdit) {
-      // A finished move/resize is a human decision about where the box goes,
-      // so it commits as a manual box.
-      if (didDragBoxRef.current) commitDrawn(boxEdit.next);
-      setBoxEdit(null);
-    }
-
-    if (currentDrawing) {
-      const start = imageToNormalized(currentDrawing.startX, currentDrawing.startY);
-      const end = imageToNormalized(currentDrawing.currentX, currentDrawing.currentY);
-      const minX = Math.min(start.x, end.x);
-      const maxX = Math.max(start.x, end.x);
-      const minY = Math.min(start.y, end.y);
-      const maxY = Math.max(start.y, end.y);
-      const threshold = 10 / (imgRef.current?.offsetWidth || 1000);
-      if (maxX - minX > threshold && maxY - minY > threshold) {
-        commitDrawn([minX, minY, maxX, maxY]);
-      } else {
-        // Too small to be a box, so it was a click: drop the selection, which
-        // is what a press on the image away from the box means.
-        setBoxSelected(false);
-      }
-      setCurrentDrawing(null);
-    }
-
-    if (isDragging) setIsDragging(false);
-  };
 
   // Clicks resolve on mouse-up, where the drag's size decides whether it was a
   // box or a deselect; this only keeps the press from escaping the editor.
   const handleClick = (e: React.MouseEvent) => e.stopPropagation();
-
-  const getCursorStyle = () => {
-    if (spaceHeld) return isDragging ? 'grabbing' : 'grab';
-    return 'crosshair';
-  };
 
   // Warm the frames either side, so an arrow press swaps to a bitmap the
   // browser has already decoded rather than to an empty <img>. Prefetching
@@ -739,23 +479,6 @@ export function LocalizeObjectEditor({
     document.addEventListener('mousedown', onPointerDown);
     return () => document.removeEventListener('mousedown', onPointerDown);
   }, [acceptOpen]);
-
-  // Non-passive so preventDefault works — the page behind must not scroll.
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container) return;
-    const onWheel = (e: WheelEvent) => {
-      e.preventDefault();
-      setTransformOrigin({ x: 50, y: 50 });
-      setZoomLevel(z => {
-        const next = Math.max(1, Math.min(4, z + (e.deltaY < 0 ? 0.2 : -0.2)));
-        if (next === 1) setPanOffset({ x: 0, y: 0 });
-        return next;
-      });
-    };
-    container.addEventListener('wheel', onWheel, { passive: false });
-    return () => container.removeEventListener('wheel', onWheel);
-  }, []);
 
   // Close = shrink back into the grid, then let onClose navigate (which
   // unmounts this component). The guard makes a second close attempt a
@@ -867,7 +590,7 @@ export function LocalizeObjectEditor({
         case 'r':
         case 'R':
           setCropView(false);
-          resetZoom();
+          resetStageZoom();
           break;
         case '?':
           setShortcutsOpen(open => !open);
@@ -898,7 +621,7 @@ export function LocalizeObjectEditor({
     boxSelected,
     acceptOpen,
     shortcutsOpen,
-    resetZoom,
+    resetStageZoom,
     requestClose,
     editable,
     isAccepting,
@@ -1084,23 +807,23 @@ export function LocalizeObjectEditor({
             selectedSmokeType={smokeType}
             objectOverlays={showOtherObjects ? objectOverlays : []}
             isDrawMode={currentDrawing !== null}
-            onBoxPointerDown={handleBoxPointerDown}
-            onHandlePointerDown={handleHandlePointerDown}
-            currentDrawing={currentDrawing}
+            onBoxPointerDown={stage.handleBoxPointerDown}
+            onHandlePointerDown={stage.handleHandlePointerDown}
+            currentDrawing={stage.currentDrawing}
             containerRef={containerRef}
             imgRef={imgRef}
-            imageInfo={imageInfo}
-            zoomLevel={zoomLevel}
-            panOffset={panOffset}
-            transformOrigin={transformOrigin}
-            isDragging={isDragging}
-            onMouseDown={handleMouseDown}
-            onMouseMove={handleMouseMove}
-            onMouseUp={handleMouseUp}
+            imageInfo={stage.imageInfo}
+            zoomLevel={stage.zoomLevel}
+            panOffset={stage.panOffset}
+            transformOrigin={stage.transformOrigin}
+            isDragging={stage.isDragging}
+            onMouseDown={stage.handleMouseDown}
+            onMouseMove={stage.handleMouseMove}
+            onMouseUp={stage.handleMouseUp}
             onClick={handleClick}
-            getCursorStyle={getCursorStyle}
-            handleImageLoad={handleImageLoad}
-            normalizedToImage={normalizedToImage}
+            getCursorStyle={stage.getCursorStyle}
+            handleImageLoad={stage.handleImageLoad}
+            normalizedToImage={stage.normalizedToImage}
             overlaysVisible
           />
