+++ frontend/src/hooks/annotation/index.ts
@@ -4,3 +4,12 @@
 
 export { useDrawingCanvas } from './useDrawingCanvas';
 export type { DrawingCanvasConfig, DrawingCanvasState } from './useDrawingCanvas';
+
+export { useBoxDrawingStage } from './useBoxDrawingStage';
+export type {
+  BoxDrawingStage,
+  BoxEditState,
+  ImageGeometry,
+  UseBoxDrawingStageParams,
+  Xyxyn,
+} from './useBoxDrawingStage';
